Testimonial Carousel moved to ind folder
