+++ frontend/src/components/TestimonialCarousel/TestimonialCarousel.css
